LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/path-objects/2d.path.rect.winding.worker-expected.txt


FAIL 2d assert_equals: Red channel of the pixel at (25, 12) expected 0 but got 255
